refactor: organize project into package
