+++ sei_aneel/__init__.py
@@ -1,0 +1,8 @@
+"""Core utilities for the SEI ANEEL automation project."""
+
+__all__ = [
+    "email_utils",
+    "log_utils",
+    "progress",
+    "ui",
+]
