+++ sei_aneel/sorteio_aneel/sorteio_aneel.py
@@ -1,9 +1,8 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
-"""Rastrea pautas publicadas pela ANEEL e dispara alertas."""
+"""Consulta sorteios na ANEEL e envia alertas por e-mail."""
 
 import os
-import sys
 import requests
 from bs4 import BeautifulSoup
 from datetime import datetime
@@ -14,32 +13,20 @@ from email.mime.multipart import MIMEMultipart
 import unicodedata
 import tempfile
 import subprocess
-from urllib.parse import urljoin
-import hashlib
+import json
 from pathlib import Path
 import html
 import logging
 import shutil
 
-# Garante que o diretório raiz do projeto esteja no ``PYTHONPATH``.
-ROOT_DIR = Path(__file__).resolve().parents[1]
-sys.path.insert(0, str(ROOT_DIR))
-
-from config import load_config, load_search_terms
-# ``email_utils`` pode não estar disponível quando o script é chamado
-# diretamente fora do repositório. Tentamos o import absoluto e, se
-# necessário, o relativo para suportar execução como pacote.
-try:  # pragma: no cover - lógica de fallback
-    from email_utils import format_html_email
-except ModuleNotFoundError:  # pragma: no cover - suporte a pacote
-    from ..email_utils import format_html_email  # type: ignore
-from log_utils import get_logger
+from ..config import load_config, load_search_terms
+from ..email_utils import format_html_email
+from ..log_utils import get_logger
 
 # Diretório de dados e arquivos de log
-DATA_DIR = os.environ.get("PAUTA_DATA_DIR", os.path.join(os.path.expanduser("~"), ".pauta_aneel"))
+DATA_DIR = os.environ.get("SORTEIO_DATA_DIR", os.path.join(os.path.expanduser("~"), ".sorteio_aneel"))
 os.makedirs(DATA_DIR, exist_ok=True)
-LOG_FILE = os.environ.get("PAUTA_LOG_FILE", os.path.join(DATA_DIR, "pauta_aneel.log"))
-HASH_RESULT_FILE = os.environ.get("HASH_RESULT_FILE", os.path.join(DATA_DIR, "ultimo_resultado_aneel_425.txt"))
+LOG_FILE = os.environ.get("SORTEIO_LOG_FILE", os.path.join(DATA_DIR, "sorteio_aneel.log"))
 
 # === Registro de data/hora de execução no log ===
 logger = get_logger(__name__, log_file=LOG_FILE)
@@ -59,8 +46,12 @@ SMTP_USER = SMTP_CONF.get("user", "")
 SMTP_PASSWORD = SMTP_CONF.get("password", "")
 EMAIL_TO = ",".join(CONFIG.get("email", {}).get("recipients", []))
 
-BASE_URL = "https://www2.aneel.gov.br/aplicacoes_liferay/noticias_area/?idAreaNoticia=425"
+BASE_URL = "https://www2.aneel.gov.br/aplicacoes_liferay/noticias_area/?idAreaNoticia=424"
 SITE_PREFIX = "https://www2.aneel.gov.br"
+
+LAST_RESULT_FILE = os.environ.get(
+    "LAST_RESULT_FILE", os.path.join(DATA_DIR, "ultimo_resultado_aneel.json")
+)
 
 # Termos de pesquisa centralizados em ``search_terms.txt``
 KEYWORDS = load_search_terms()
@@ -81,7 +72,6 @@ def palavra_chave_no_texto(texto, palavras_chave):
 
 
 def should_notify(items, erro=False):
-    """Retorna True se deve enviar notificação."""
     return erro or not items
 
 
@@ -92,11 +82,27 @@ def parse_date(date_str):
     except Exception:
         return None
 
+def ler_ultimo_resultado():
+    try:
+        with open(LAST_RESULT_FILE, "r") as f:
+            return json.load(f)
+    except Exception:
+        return {"data_encontrada": None, "items": []}
+
+def salvar_ultimo_resultado(data_encontrada, items):
+    try:
+        with open(LAST_RESULT_FILE, "w") as f:
+            json.dump({
+                "data_encontrada": data_encontrada,
+                "items": items
+            }, f)
+    except Exception as e:
+        registrar_log(f"Erro ao salvar último resultado: {e}")
+
 def find_nearest_date_link(target_date=None):
     headers = {"User-Agent": "Mozilla/5.0"}
     response = requests.get(BASE_URL, timeout=30, headers=headers)
     if not response.ok:
-        logger.error(f"Erro ao acessar ANEEL: {response.status_code}")
         registrar_log(f"Erro ao acessar ANEEL: {response.status_code}")
         return None, None, None
     soup = BeautifulSoup(response.text, "html.parser")
@@ -121,7 +127,6 @@ def find_nearest_date_link(target_date=None):
         if isinstance(target_date, str):
             target_date = parse_date(target_date)
         if not target_date:
-            logger.error("Data inválida informada.")
             registrar_log("Data inválida informada.")
             return None, None, None
     for date_found, link_found, link_text in found_rows:
@@ -135,92 +140,60 @@ def extract_items_from_tr(url):
     response = requests.get(url, timeout=30, headers=headers)
     response.encoding = response.apparent_encoding
     if not response.ok:
-        logger.error(f"Erro ao acessar documento: {response.status_code}")
         registrar_log(f"Erro ao acessar documento: {response.status_code}")
         return []
     soup = BeautifulSoup(response.text, "html.parser")
-
-    if soup.body:
-        content_elements = []
-        for tag in soup.body.descendants:
-            if tag.name in ['p', 'div', 'li']:
-                content_elements.append(tag)
-    else:
-        content_elements = soup.find_all(['p', 'div', 'li'])
-
-    idx = 0
+    all_tr = soup.find_all("tr")
+    main_tr = max(all_tr, key=lambda tr: len(tr.get_text()), default=None)
+    if not main_tr:
+        registrar_log("Nenhum <tr> relevante encontrado.")
+        return []
+    full_text = main_tr.get_text(separator=" ", strip=True)
+    marker_regex = re.compile(r'(?<!\d)(\d{1,3})\. Processo: \d{5,}\.')
+    matches = list(marker_regex.finditer(full_text))
+    items = []
+    for idx, match in enumerate(matches):
+        start = match.start()
+        end = matches[idx+1].start() if idx+1 < len(matches) else len(full_text)
+        item = full_text[start:end].strip()
+        items.append(item)
     final_results = []
-    while idx < len(content_elements):
-        el = content_elements[idx]
-        texto = el.get_text(separator=" ", strip=True)
-        m_proc = re.match(r'^\d+\.\s*Processo\s*:\s*[\d\.]+/\d{4}-\d{2}', texto)
-        if m_proc:
-            bloco_texto = texto
-            bloco_idx_start = idx
-            processo_numero = None
-            processo_numero_pdf = ""
-            m_proc_n = re.search(r'Processo\s*:\s*([\d\.]+/\d{4}-\d{2})', bloco_texto)
-            if m_proc_n:
-                processo_numero = m_proc_n.group(1)
-                processo_numero_pdf = re.sub(r'[\./-]', '', processo_numero)
-            pdfs = []
-            doc_count = 0
-            idx_next = idx + 1
-            bloco_extra_text = ""
-            while idx_next < len(content_elements):
-                texto_next = content_elements[idx_next].get_text(separator=" ", strip=True)
-                if re.match(r'^\d+\.\s*Processo\s*:\s*[\d\.]+/\d{4}-\d{2}', texto_next):
-                    break
-                for a in content_elements[idx_next].find_all('a', href=True):
-                    href = a['href']
-                    if 'pdf' in href.lower():
-                        doc_count += 1
-                        pdf_url = urljoin(url, href)
-                        doc_index = f" - doc{doc_count}" if doc_count > 1 else ""
-                        pdf_filename = f"Minutas de voto e ato - {processo_numero_pdf}{doc_index}.pdf"
-                        pdfs.append({"pdf_url": pdf_url, "pdf_filename": pdf_filename})
-                bloco_extra_text += " " + texto_next
-                idx_next += 1
-            bloco_texto_full = bloco_texto + bloco_extra_text
-            if palavra_chave_no_texto(bloco_texto_full, KEYWORDS):
-                final_results.append({
-                    "text": bloco_texto_full.strip(),
-                    "processo_numero": processo_numero,
-                    "processo_numero_pdf": processo_numero_pdf,
-                    "pdfs": pdfs
-                })
-            idx = idx_next
-        else:
-            idx += 1
+    for item in items:
+        if palavra_chave_no_texto(item, KEYWORDS):
+            item_clean = re.sub(r'\s+', ' ', item).strip()
+            final_results.append(item_clean)
     return final_results
 
 def gerar_pdf_da_pagina(url, pdf_file):
     try:
         if not shutil.which("wkhtmltopdf"):
-            logger.error("wkhtmltopdf não encontrado")
+            registrar_log("wkhtmltopdf não encontrado")
             return False
-        env = os.environ.copy()
-        env.setdefault("XDG_RUNTIME_DIR", "/tmp")
+        headers = {'User-Agent': 'Mozilla/5.0'}
+        html = requests.get(url, headers=headers, timeout=30).text
+        with tempfile.NamedTemporaryFile(delete=False, suffix=".html") as tmp_html:
+            tmp_html.write(html.encode("utf-8"))
+            html_path = tmp_html.name
         result = subprocess.run(
-            ["wkhtmltopdf", "--quiet", "--load-error-handling", "ignore", url, pdf_file],
+            ["wkhtmltopdf", "--quiet", html_path, pdf_file],
             stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-            env=env
+            stderr=subprocess.PIPE
         )
         if result.returncode != 0:
-            logger.error("Erro ao gerar PDF (wkhtmltopdf): %s", result.stderr.decode())
             registrar_log(f"Erro ao gerar PDF (wkhtmltopdf): {result.stderr.decode()}")
+            os.remove(html_path)
             return False
+        os.remove(html_path)
         return True
     except Exception as e:
-        logger.error(f"Erro ao gerar PDF: {e}")
         registrar_log(f"Erro ao gerar PDF: {e}")
         return False
 
 def send_email(subject, body_plain, body_html, pdf_path=None):
     msg = MIMEMultipart()
     msg["From"] = SMTP_USER
-    msg["To"] = EMAIL_TO
+    destinatarios = [e.strip() for e in EMAIL_TO.replace(';', ',').split(',') if e.strip()]
+    msg["To"] = ", ".join(destinatarios)
     msg["Subject"] = subject
     pdf_attached = False
     if pdf_path and os.path.exists(pdf_path) and os.path.getsize(pdf_path) > 0:
@@ -234,7 +207,7 @@ def send_email(subject, body_plain, body_html, pdf_path=None):
             msg.attach(part)
             pdf_attached = True
     else:
-        logger.warning("PDF não gerado ou está vazio, não será anexado.")
+        registrar_log("PDF não gerado ou está vazio, não será anexado.")
         aviso = (
             "\n\nATENÇÃO: Não foi possível anexar o PDF da página, pois ocorreu um erro na geração do arquivo.\n"
         )
@@ -249,29 +222,13 @@ def send_email(subject, body_plain, body_html, pdf_path=None):
     try:
         with smtplib.SMTP(SMTP_SERVER, SMTP_PORT) as server:
             server.login(SMTP_USER, SMTP_PASSWORD)
-            to_emails = [email.strip() for email in EMAIL_TO.split(",")]
-            server.sendmail(SMTP_USER, to_emails, msg.as_string())
+            server.sendmail(SMTP_USER, destinatarios, msg.as_string())
         logger.info("E-mail enviado com sucesso.")
         registrar_log(f"E-mail enviado para {EMAIL_TO}")
-        registrar_log("Corpo do e-mail:\n" + body_plain + "\n---")
+        registrar_log("Corpo do e-mail:\n" + body_plain)
     except Exception as e:
         logger.error(f"Erro ao enviar e-mail: {e}")
         registrar_log(f"Falha ao enviar e-mail: {e}")
-
-def hash_do_resultado(itens):
-    texto = "\n".join([item['text'] for item in itens])
-    return hashlib.sha256(texto.encode('utf-8')).hexdigest()
-
-def ler_ultimo_hash(arquivo):
-    try:
-        with open(arquivo, "r") as f:
-            return f.read().strip()
-    except FileNotFoundError:
-        return None
-
-def salvar_hash(arquivo, hash_str):
-    with open(arquivo, "w") as f:
-        f.write(hash_str)
 
 def main():
     if len(sys.argv) > 1:
@@ -280,24 +237,37 @@ def main():
     else:
         data_pesquisa = None
         execucao_manual = False
+
     hoje = datetime.now()
     hoje_str = hoje.strftime("%d/%m/%Y")
     url, link_text, data_encontrada = find_nearest_date_link(data_pesquisa or hoje_str)
     pdf_path = None
+
     if not url or not data_encontrada:
         logger.info("Nenhum link associado à data encontrada.")
-        subject = f"{hoje_str} Busca Pauta ANEEL - Nenhuma data encontrada"
-        body = "Nao encontrada pauta para data indicada! Atenciosamente, Ary Abdo!"
+        subject = f"{hoje_str} Busca Sorteio ANEEL - Nenhuma data encontrada"
+        body = "Nao encontrado sorteio para data indicada! Atenciosamente, Ary Abdo!"
         content_html = (
             "<div class=\"section\">"
-            "<p>Nao encontrada pauta para data indicada!</p>"
+            "<p>Nao encontrado sorteio para data indicada!</p>"
             "<p>Atenciosamente,<br>Ary Abdo</p>"
             "</div>"
         )
-        body_html = format_html_email("Pauta da Próxima Reunião ANEEL", content_html)
+        body_html = format_html_email("Sorteio ANEEL", content_html)
         if should_notify([], True):
             send_email(subject, body, body_html)
         registrar_log("Nenhum link associado à data encontrada.")
+        return
+
+    items = extract_items_from_tr(url)
+    ultimo_resultado = ler_ultimo_resultado()
+    items_anteriores = ultimo_resultado.get("items", [])
+    data_encontrada_anterior = ultimo_resultado.get("data_encontrada")
+
+    notificar = should_notify(items)
+    if not execucao_manual and not notificar and items == items_anteriores:
+        logger.info("Nenhuma atualização nos itens encontrados.")
+        registrar_log("Nenhuma atualização nos itens encontrados.")
         return
 
     if url:
@@ -307,63 +277,40 @@ def main():
         if not sucesso_pdf:
             logger.warning("PDF não gerado, mas tentando anexar se existir.")
 
-    items = extract_items_from_tr(url)
-    subject = f"{hoje_str} Busca Pauta ANEEL - {data_encontrada} - {link_text}"
-
+    subject = f"{hoje_str} Busca Sorteio ANEEL - {data_encontrada} - {link_text}"
     if items:
-        body = "Foram encontrados os processos listados abaixo na pauta da próxima reuniao da ANEEL:\n\n"
+        body = (
+            "Foram encontrados os processos listados abaixo no sorteio realizado pela ANEEL:\n\n"
+            + "\n\n".join(items)
+            + "\n\nAtenciosamente,\nAry Abdo"
+        )
         content_html = (
             "<div class=\"section\">"
             f"<h3>📋 Processos Encontrados ({len(items)})</h3>"
             "<ul>"
         )
         for item in items:
-            body += item["text"] + "\n"
-            item_html = html.escape(item["text"])
-            content_html += f"<li class=\"item\">{item_html}"
-            registrar_log("Processo encontrado:\n" + item["text"])
-            if item["pdfs"]:
-                content_html += "<ul>"
-                for pdf in item["pdfs"]:
-                    body += f"(Acesse o documento PDF: {pdf['pdf_url']})\n"
-                    link_html = html.escape(pdf['pdf_filename'])
-                    content_html += f"<li><a href='{pdf['pdf_url']}'>{link_html}</a></li>"
-                content_html += "</ul>"
-                body += "\n"
-            else:
-                body += "Documentos nao disponibilizados.\n\n"
-                content_html += "<p>Documentos nao disponibilizados.</p>"
-            content_html += "</li>"
-        body += "\nAtenciosamente,\nAry Abdo"
+            content_html += f"<li class=\"item\">{html.escape(item)}</li>"
+            registrar_log(f"Processo encontrado: {item}")
         content_html += "</ul><p>Atenciosamente,<br>Ary Abdo</p></div>"
-        body_html = format_html_email("Pauta da Próxima Reunião ANEEL", content_html)
+        body_html = format_html_email("Sorteio ANEEL", content_html)
     else:
-        body = "Ola! Nao foram encontrados processos listados na pauta na data de pesquisa!\n\nAtenciosamente, Ary Abdo!"
+        body = "Ola! Nao foram encontrados processos sorteados na data de pesquisa!\n\nAtenciosamente, Ary Abdo!"
         content_html = (
             "<div class=\"section\">"
-            "<p>Ola! Nao foram encontrados processos listados na pauta na data de pesquisa!</p>"
+            "<p>Ola! Nao foram encontrados processos sorteados na data de pesquisa!</p>"
             "<p>Atenciosamente,<br>Ary Abdo</p>"
             "</div>"
         )
-        body_html = format_html_email("Pauta da Próxima Reunião ANEEL", content_html)
+        body_html = format_html_email("Sorteio ANEEL", content_html)
         registrar_log("Nenhum processo relevante encontrado.")
 
     logger.info("Itens relevantes encontrados:" if items else "Nenhum item relevante encontrado.")
     logger.info(body)
-
-    hash_arquivo = HASH_RESULT_FILE
-    hash_atual = hash_do_resultado(items)
-    hash_antigo = ler_ultimo_hash(hash_arquivo)
-    notificar = should_notify(items)
-
-    if execucao_manual or notificar or (hash_atual != hash_antigo):
+    if execucao_manual or notificar or items != items_anteriores:
         send_email(subject, body, body_html, pdf_path)
-        salvar_hash(hash_arquivo, hash_atual)
-        registrar_log("Conteúdo alterado/enviado. Hash atualizado.")
-    else:
-        logger.info("Nenhuma atualização no conteúdo da pauta ANEEL.")
-        registrar_log("Nenhuma atualização na data da pauta ANEEL.")
-
+    if not execucao_manual:
+        salvar_ultimo_resultado(data_encontrada, items)
     if pdf_path and os.path.exists(pdf_path):
         os.remove(pdf_path)
